LayoutTests/inspector/console/js-isLikelyStackTrace-expected.txt

Test stack trace detection heuristic.


== Running test suite: WI.StackTrace.isLikelyStackTrace
-- Running test case: notStacktrace
PASS: Should NOT be a stacktrace.

-- Running test case: typeErrorWrap
FAIL: Should be a stacktrace.
    Expected: truthy
    Actual: false

-- Running test case: getAnonymousFunctionError
PASS: Should be a stacktrace.

-- Running test case: testWithNativeCallInBetween
FAIL: Should be a stacktrace.
    Expected: truthy
    Actual: false

-- Running test case: testFormattedStrings
PASS: "video:1:2" should not be a stacktrace.
PASS: "video/mp4:1:2" should not be a stacktrace.
PASS: "video/mp4 : 11:22:33-44:55:66" should not be a stacktrace.
PASS: "http://video:1:2" should not be a stacktrace.
PASS: "http://video/mp4:1:2" should not be a stacktrace.
PASS: "http://video/mp4 : 11:22:33-44:55:66" should not be a stacktrace.
